# Vectorize benchmark functions