--- a/examples/benchmarks/schaffer_f7/globmin
+++ b/examples/benchmarks/schaffer_f7/globmin
@@ -1,0 +1,38 @@
+<PARAMETERSFOR>
+schafferf7bench:25D
+</PARAMETERSFOR>
+<PARAMETERFITNESS>
+ 0.0
+</PARAMETERFITNESS>
+<PARAMETERS>
+<ATOM>
+XX
+</ATOM>
+<VALUES>
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+ 0.0
+</VALUES>
+</PARAMETERS>