--- a/src/org/ogolem/adaptive/BenchAckley.java
+++ b/src/org/ogolem/adaptive/BenchAckley.java
@@ -41,6 +41,9 @@ package org.ogolem.adaptive;
 import static java.lang.Math.*;
 
 import java.util.ArrayList;
+import jdk.incubator.vector.DoubleVector;
+import jdk.incubator.vector.VectorOperators;
+import jdk.incubator.vector.VectorSpecies;
 import org.ogolem.core.BondInfo;
 import org.ogolem.core.CartesianCoordinates;
 
@@ -48,26 +51,26 @@ import org.ogolem.core.CartesianCoordinates;
  * Ackleys benchmark function in nD. Global minimum: f(x)=0, x(i)=0.0.
  *
  * @author Johannes Dieterich
- * @version 2020-12-29
+ * @version 2021-07-21
  */
-final class BenchAckley extends AbstractAdaptivable {
+public final class BenchAckley extends AbstractAdaptivable {
 
   private static final long serialVersionUID = (long) 20150727;
   private static final boolean DEBUG = false;
   private static final double TWOPI = PI * 2;
-  private final boolean correctGrad;
+  private static final VectorSpecies<Double> SPECIES = DoubleVector.SPECIES_PREFERRED;
+
   private final int iDims;
   private final double dimsInv;
 
-  BenchAckley(final int dims, final boolean corrGrad) {
+  public BenchAckley(final int dims) {
     this.iDims = dims;
     this.dimsInv = 1.0 / dims;
-    this.correctGrad = corrGrad;
   }
 
   @Override
   public BenchAckley copy() {
-    return new BenchAckley(iDims, correctGrad);
+    return new BenchAckley(iDims);
   }
 
   @Override
@@ -78,18 +81,46 @@ final class BenchAckley extends AbstractAdaptivable {
       final BondInfo bonds) {
 
     final double[] p = params.getAllParamters();
-    final int no = p.length;
 
-    double a = 0.0;
-    double b = 0.0;
-    for (int i = 0; i < no; i++) {
+    final int upperBound = SPECIES.loopBound(iDims);
+    final var vCutVal = DoubleVector.broadcast(SPECIES, 32.768);
+    final var vCutCons = DoubleVector.broadcast(SPECIES, FixedValues.NONCONVERGEDENERGY);
+    final var vTwoPI = DoubleVector.broadcast(SPECIES, 2 * Math.PI);
+
+    var vA = DoubleVector.broadcast(SPECIES, 0.0);
+    var vB = DoubleVector.broadcast(SPECIES, 0.0);
+    int i = 0;
+    for (; i < upperBound; i += SPECIES.length()) {
+      final var vPI = DoubleVector.fromArray(SPECIES, p, i);
+      // first check if we need the cutoff
+      final var vPIAbs = vPI.abs();
+      final var vCutMask = vPIAbs.compare(VectorOperators.GT, vCutVal);
+      final var anyCut = vCutMask.anyTrue();
+      if (anyCut) {
+        final var vPISqUn = vPI.mul(vPI);
+        final var vPISq = vPISqUn.blend(vCutCons, vCutMask);
+        vA = vPISq.add(vA);
+
+        final var vBTemUn = vPI.mul(vTwoPI).lanewise(VectorOperators.COS);
+        final var vBTem = vBTemUn.blend(vCutCons, vCutMask);
+        vB = vBTem.add(vB);
+      } else {
+        vA = vPI.fma(vPI, vA);
+        final var vBTem = vPI.mul(vTwoPI).lanewise(VectorOperators.COS);
+        vB = vBTem.add(vB);
+      }
+    }
+
+    double a = vA.reduceLanes(VectorOperators.ADD);
+    double b = vB.reduceLanes(VectorOperators.ADD);
+    for (; i < iDims; i++) {
       final double pi = p[i];
       if (pi > 32.768 || pi < -32.768) {
         a += FixedValues.NONCONVERGEDENERGY;
         b += FixedValues.NONCONVERGEDENERGY;
         continue;
       }
-      a += pi * pi;
+      a = Math.fma(pi, pi, a);
       b += cos(TWOPI * pi);
     }
 
@@ -108,11 +139,47 @@ final class BenchAckley extends AbstractAdaptivable {
 
     final double[] p = params.getAllParamters();
     assert (p.length == params.getNumberOfParamters());
-    final int no = params.getNumberOfParamters();
 
-    double a = 0.0;
-    double b = 0.0;
-    for (int i = 0; i < no; i++) {
+    final int upperBound = SPECIES.loopBound(iDims);
+    final var vCutVal = DoubleVector.broadcast(SPECIES, 32.768);
+    final var vCutCons = DoubleVector.broadcast(SPECIES, FixedValues.NONCONVERGEDENERGY);
+    final var vTwoPI = DoubleVector.broadcast(SPECIES, 2 * Math.PI);
+
+    var vA = DoubleVector.broadcast(SPECIES, 0.0);
+    var vB = DoubleVector.broadcast(SPECIES, 0.0);
+    int i = 0;
+    for (; i < upperBound; i += SPECIES.length()) {
+      final var vPI = DoubleVector.fromArray(SPECIES, p, i);
+      // first check if we need the cutoff
+      final var vPIAbs = vPI.abs();
+      final var vCutMask = vPIAbs.compare(VectorOperators.GT, vCutVal);
+      final var anyCut = vCutMask.anyTrue();
+      if (anyCut) {
+        final var vPISqUn = vPI.mul(vPI);
+        final var vPISq = vPISqUn.blend(vCutCons, vCutMask);
+        vA = vPISq.add(vA);
+
+        final var vTrigi = vPI.mul(vTwoPI);
+        final var vBTemUn = vTrigi.lanewise(VectorOperators.COS);
+        final var vBTem = vBTemUn.blend(vCutCons, vCutMask);
+        vB = vBTem.add(vB);
+
+        final var vGradTmp = vTrigi.lanewise(VectorOperators.SIN).mul(vTwoPI);
+        vGradTmp.intoArray(daGrad, i);
+
+      } else {
+        vA = vPI.fma(vPI, vA);
+        final var vTrigi = vPI.mul(vTwoPI);
+        final var vBTem = vTrigi.lanewise(VectorOperators.COS);
+        vB = vBTem.add(vB);
+        final var vGradTmp = vTrigi.lanewise(VectorOperators.SIN).mul(vTwoPI);
+        vGradTmp.intoArray(daGrad, i);
+      }
+    }
+
+    double a = vA.reduceLanes(VectorOperators.ADD);
+    double b = vB.reduceLanes(VectorOperators.ADD);
+    for (; i < iDims; i++) {
       /*
        * done using wolframalpha :-)
        */
@@ -130,38 +197,41 @@ final class BenchAckley extends AbstractAdaptivable {
         continue;
       }
 
-      final double pSq = pi * pi;
-      a += pSq;
+      a = Math.fma(pi, pi, a);
       final double trigi = TWOPI * pi;
       final double bi = cos(trigi);
       b += bi;
 
-      if (correctGrad) {
-        daGrad[i] = TWOPI * sin(trigi); // we use this as a cache
-      } else {
-        final double pSqDims = sqrt(pSq * dimsInv);
-
-        final double firstNum = 4 * exp(-0.2 * pSqDims) * pi;
-        final double firstDenom = pSqDims * iDims;
-        final double firstTerm = firstNum / firstDenom;
-        final double secNum = TWOPI * sin(trigi) * exp(dimsInv * bi);
-        final double secTerm = secNum * dimsInv;
-
-        daGrad[i] = firstTerm + secTerm;
-      }
+      daGrad[i] = TWOPI * sin(trigi); // we use this as a cache
     }
 
     final double t1 = exp(b * dimsInv);
     final double t2 = sqrt(a * dimsInv);
-    final double t2Inv = 1 / (iDims * t2);
+    final double t2Inv = (Math.abs(t2) < 1e-10) ? 0.0 : 1 / (iDims * t2);
     final double t3 = exp(-0.2 * t2);
 
-    if (correctGrad) {
-      for (int i = 0; i < no; i++) {
-        final double secNum = daGrad[i] * t1;
-        final double secTerm = secNum * dimsInv;
-        daGrad[i] = 4 * p[i] * t3 * t2Inv + secTerm;
-      }
+    final var vT1 = DoubleVector.broadcast(SPECIES, t1);
+    final var vT2Inv = DoubleVector.broadcast(SPECIES, t2Inv);
+    final var vT3 = DoubleVector.broadcast(SPECIES, t3);
+    final var vFour = DoubleVector.broadcast(SPECIES, 4);
+    final var vDimsInv = DoubleVector.broadcast(SPECIES, dimsInv);
+
+    int x = 0;
+    for (; x < upperBound; x += SPECIES.length()) {
+      final var vGradLoad = DoubleVector.fromArray(SPECIES, daGrad, x);
+      final var vSecNum = vGradLoad.mul(vT1);
+      final var vSecTerm = vSecNum.mul(vDimsInv);
+      final var vP = DoubleVector.fromArray(SPECIES, p, x);
+      final var vGrad = vP.mul(vFour).mul(vT3).mul(vT2Inv).add(vSecTerm);
+
+      vGrad.intoArray(daGrad, x);
+    }
+
+    for (; x < iDims; x++) {
+      final double secNum = daGrad[x] * t1;
+      final double secTerm = secNum * dimsInv;
+
+      daGrad[x] = 4 * p[x] * t3 * t2Inv + secTerm;
     }
 
     final double e = -20 * t3 - t1 + 20 + E;
@@ -173,8 +243,8 @@ final class BenchAckley extends AbstractAdaptivable {
           NumericalGradients.calculateParamGrad(cartes, params, this, geomID, bonds, numGrad);
       System.out.println("DEBUG OUTPUT FOR NUMGRAD");
       System.out.println("Func " + numFunc + "\t" + e + "\t" + (numFunc - e));
-      for (int i = 0; i < daGrad.length; i++) {
-        System.out.println(numGrad[i] + "\t" + daGrad[i] + "\t" + (numGrad[i] - daGrad[i]));
+      for (int y = 0; y < daGrad.length; y++) {
+        System.out.println(numGrad[y] + "\t" + daGrad[y] + "\t" + (numGrad[y] - daGrad[y]));
       }
     }