Also add unit tests, add microbenchmarks, add macrobenchmarks.

From @bernd-hartke on Linux/amd64 w/ Java16 vs Java17 EA (which contains SVML vectorized trig functions):

here is java16:

WARNING: Using incubator modules: jdk.incubator.vector
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Ackley function benchmark in 250D took 297 +/- 20 ms per macro iteration
                         or 336323 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Ackley gradient benchmark in 250D took 538 +/- 6 ms per macro iteration
                         or 185989 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Lunacek function benchmark in 250D took 289 +/- 2 ms per macro iteration
                         or 346420 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Lunacek gradient benchmark in 250D took 556 +/- 5 ms per macro iteration
                         or 179748 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schaffer F7 function benchmark in 250D took 741 +/- 1 ms per macro iteration
                         or 135014 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schaffer F7 gradient benchmark in 250D took 1563 +/- 2 ms per macro iteration
                         or 63980 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Rastrigin function benchmark in 250D took 280 +/- 1 ms per macro iteration
                         or 356718 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Rastrigin gradient benchmark in 250D took 535 +/- 2 ms per macro iteration
                         or 186800 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schwefel function benchmark in 250D took 281 +/- 1 ms per macro iteration
                         or 355450 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schwefel gradient benchmark in 250D took 547 +/- 2 ms per macro iteration
                         or 182927 calls per second on average

vs. java17:

WARNING: Using incubator modules: jdk.incubator.vector
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Ackley function benchmark in 250D took 82 +/- 2 ms per macro iteration
                         or 1219512 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Ackley gradient benchmark in 250D took 148 +/- 19 ms per macro iteration
                         or 675676 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Lunacek function benchmark in 250D took 70 +/- 5 ms per macro iteration
                         or 1421801 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Lunacek gradient benchmark in 250D took 143 +/- 4 ms per macro iteration
                         or 700935 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schaffer F7 function benchmark in 250D took 575 +/- 3 ms per macro iteration
                         or 173812 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schaffer F7 gradient benchmark in 250D took 1196 +/- 3 ms per macro iteration
                         or 83612 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Rastrigin function benchmark in 250D took 64 +/- 1 ms per macro iteration
                         or 1554404 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Rastrigin gradient benchmark in 250D took 127 +/- 0 ms per macro iteration
                         or 787402 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schwefel function benchmark in 250D took 94 +/- 0 ms per macro iteration
                         or 1063830 calls per second on average
[main] INFO org.ogolem.microbenchmarks.MainMicroBenchmarks - Benchmark Schwefel gradient benchmark in 250D took 162 +/- 1 ms per macro iteration
                         or 617284 calls per second on average

Basically, things will get much faster on supported platforms w/ vectorized trigs.